+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Change Log
 - added `-l` `--language` command line options and `language` option in config file
 - added `-j` `--first` command line option to open the first result and exit
 - added `--lucky` command line option to open a random result and exit
+- added `week` option to `--time-range` option and enabled `d`, `w`, `m`, `y` as short codes